Corrigido bug ao remover usuário com avaliações e comentários

+++ src/main/java/br/unesp/grupo5/trabalhofinal/service/UploadService.java
@@ -89,16 +89,18 @@ public class UploadService {
     public FileSystemResource getConteudoVideo(String name) {
         return readFile(conteudoVideo.toString(), name);
     }
-    
+
     public void deleteFile(String dir, String name) {
-        Path location = Paths.get(dir, name);
-        try {
-            Files.delete(location);
-        } catch (IOException ex) {
-            System.out.println("Arquivo nao encontrado: " + location.toString());
+        if (dir != null && name != null) {
+            Path location = Paths.get(dir, name);
+            try {
+                Files.delete(location);
+            } catch (Exception ex) {
+                System.out.println("Arquivo nao encontrado: " + location.toString());
+            }
         }
     }
-    
+
     public void deleteSerieThumb(String name) {
         deleteFile(seriesThumb.toString(), name);
     }
